added new doc for jill, inlcuded delete-disconnect function
