+++ __init__.py
@@ -1,2 +1,3 @@
 from .pyMAP import tools,data,plt,tof
 from pyMAP import bowPy
+from .pyMAP.jill import jill
